refactor: Move some functions from Lattice impls to VoxelSpaceBase
