+++ ecell4/core/LatticeSpaceVectorImpl.hpp
@@ -39,19 +39,7 @@ public:
 
     Integer num_species() const;
 
-    virtual Real get_value(const Species& sp) const
-    {
-        return static_cast<Real>(num_molecules(sp));
-    }
-
-    virtual Real get_value_exact(const Species& sp) const
-    {
-        return static_cast<Real>(num_molecules_exact(sp));
-    }
-
-    bool has_species(const Species& sp) const;
     // bool has_species_exact(const Species& sp) const;
-    virtual bool has_voxel(const ParticleID& pid) const;
 
     virtual bool remove_voxel(const ParticleID& pid);
     virtual bool remove_voxel(const coordinate_type& coord);
@@ -73,32 +61,15 @@ public:
     virtual std::pair<ParticleID, Voxel> get_voxel(const ParticleID& pid) const;
     virtual std::pair<ParticleID, Voxel> get_voxel_at(const coordinate_type& coord) const;
 
-    virtual Integer num_voxels_exact(const Species& sp) const;
-    virtual Integer num_voxels(const Species& sp) const;
-    virtual Integer num_voxels() const;
-
-    virtual Integer num_molecules(const Species& sp) const; //XXX:
-
     // virtual void update_voxel(const Voxel& v);
     virtual bool update_voxel(const ParticleID& pid, const Voxel& v);
 
     bool add_voxels(const Species species, std::vector<std::pair<ParticleID, coordinate_type> > voxels);
 
-    std::vector<Species> list_species() const;
     const Species& find_species(std::string name) const;
     std::vector<coordinate_type> list_coords(const Species& sp) const;
     std::vector<coordinate_type> list_coords_exact(const Species& sp) const;
 
-    virtual bool has_molecule_pool(const Species& sp) const
-    {
-        molecule_pool_map_type::const_iterator itr(molecule_pools_.find(sp));
-        return (itr != molecule_pools_.end());
-    }
-
-    virtual VoxelPool* find_voxel_pool(const Species& sp);
-    virtual const VoxelPool* find_voxel_pool(const Species& sp) const;
-    virtual MoleculePool* find_molecule_pool(const Species& sp);
-    virtual const MoleculePool* find_molecule_pool(const Species& sp) const;
     // VoxelPool* find_voxel_pool(const std::string name);
     virtual VoxelPool* get_voxel_pool_at(const coordinate_type& coord) const;
