+++ ecell4/core/OffLatticeSpace.cpp
@@ -414,35 +414,6 @@ OffLatticeSpace::get_neighbor_boundary(const coordinate_type& coord, const Integ
     return get_neighbor(coord, nrand);
 }
 
-Integer OffLatticeSpace::num_molecules(const Species& sp) const
-{
-    Integer count(0);
-    SpeciesExpressionMatcher sexp(sp);
-
-    for (voxel_pool_map_type::const_iterator itr(voxel_pools_.begin());
-         itr != voxel_pools_.end(); ++itr)
-    {
-        const Integer cnt(sexp.count((*itr).first));
-        if (cnt > 0)
-        {
-            const boost::shared_ptr<VoxelPool>& vp((*itr).second);
-            count += count_voxels(vp) * cnt;
-        }
-    }
-
-    for (molecule_pool_map_type::const_iterator itr(molecule_pools_.begin());
-         itr != molecule_pools_.end(); ++itr)
-    {
-        const Integer cnt(sexp.count((*itr).first));
-        if (cnt > 0)
-        {
-            const boost::shared_ptr<MoleculePool>& vp((*itr).second);
-            count += vp->size() * cnt;
-        }
-    }
-    return count;
-}
-
 Real3 OffLatticeSpace::actual_lengths() const
 {
     throw NotSupported("OffLatticeSpace::actual_lengths() is not supported.");
